# api/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "Rest接口",
    "version": "1.0",
    "title": "高并发自动化测试平台后端服务集成Swagger"
  },
...
  "paths": {
    "/config/add": {
      "post": {
        "tags": [
          "配置管理"
        ],
        "summary": "新增",
        "operationId": "addConfigUsingPOST",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "configParam",
            "description": "configParam",
            "required": true,
            "schema": {
              "$ref": "#/definitions/ConfigParam",
              "originalRef": "ConfigParam"
            }
          },
          {
            "name": "token",
            "in": "header",
            "description": "用户token认证",
            "required": false,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Response«long»",
              "originalRef": "Response«long»"
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "deprecated": false
      }
    },
    "/config/delete/{id}": {
      "get": {
        "tags": [
          "配置管理"
        ],
        "summary": "删除",
        "operationId": "deleteConfigUsingGET",
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "id",
            "required": true,
            "type": "integer",
            "format": "int64"
          },
          {
            "name": "token",
            "in": "header",
            "description": "用户token认证",
            "required": false,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Response«boolean»",
              "originalRef": "Response«boolean»"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "deprecated": false
      }
    },
    "/config/list": {
      "get": {
        "tags": [
          "配置管理"
        ],
        "summary": "分页查询",
        "operationId": "getConfigListUsingGET",
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "configKey",
            "in": "query",
            "description": "配置字段",
            "required": false,
            "type": "string"
          },
          {
            "name": "page",
            "in": "query",
            "description": "页码 第一页:1",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "name": "size",
            "in": "query",
            "description": "每页数量",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "name": "token",
            "in": "header",
            "description": "用户token认证",
            "required": false,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Response«PageVO«ConfigVO»»",
              "originalRef": "Response«PageVO«ConfigVO»»"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "deprecated": false
      }
    },
    "/config/update/{id}": {
      "post": {
        "tags": [
          "配置管理"
        ],
        "summary": "修改",
        "operationId": "updateConfigUsingPOST",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "configParam",
            "description": "configParam",
            "required": true,
            "schema": {
              "$ref": "#/definitions/ConfigParam",
              "originalRef": "ConfigParam"
            }
          },
          {
            "name": "id",
            "in": "path",
            "description": "id",
            "required": true,
            "type": "integer",
            "format": "int64"
          },
          {
            "name": "token",
            "in": "header",
            "description": "用户token认证",
            "required": false,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Response«boolean»",
              "originalRef": "Response«boolean»"
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "deprecated": false
      }
    },
...
  },
  "definitions": {
...
    "ConfigParam": {
      "type": "object",
      "required": [
        "configKey",
        "configValue"
      ],
      "properties": {
        "configKey": {
          "type": "string",
          "description": "配置字段"
        },
        "configValue": {
          "type": "string",
          "description": "字段值"
        },
        "description": {
          "type": "string",
          "description": "字段描述"
        }
      },
      "title": "ConfigParam"
    },
    "ConfigVO": {
      "type": "object",
      "properties": {
        "configKey": {
          "type": "string",
          "description": "配置字段"
        },
        "configValue": {
          "type": "string",
          "description": "字段值"
        },
        "createTime": {
          "type": "string",
          "format": "date-time"
        },
        "creator": {
          "type": "string"
        },
        "creatorId": {
          "type": "integer",
          "format": "int64"
        },
        "description": {
          "type": "string",
          "description": "字段描述"
        },
        "id": {
          "type": "integer",
          "format": "int64"
        },
        "modifier": {
          "type": "string"
        },
        "modifierId": {
          "type": "integer",
          "format": "int64"
        },
        "modifyTime": {
          "type": "string",
          "format": "date-time"
        }
      },
      "title": "ConfigVO"
    },
...
    "PageVO«ConfigVO»": {
      "type": "object",
      "properties": {
        "list": {
          "type": "array",
          "description": "列表",
          "items": {
            "$ref": "#/definitions/ConfigVO",
            "originalRef": "ConfigVO"
          }
        },
        "page": {
          "type": "integer",
          "format": "int32",
          "description": "页码"
        },
        "size": {
          "type": "integer",
          "format": "int32",
          "description": "每页数量"
        },
        "total": {
          "type": "integer",
          "format": "int32",
          "description": "总数"
        }
      },
      "title": "PageVO«ConfigVO»"
    },
...
    "Response«PageVO«ConfigVO»»": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "currentTime": {
          "type": "string",
          "format": "date-time"
        },
        "data": {
          "$ref": "#/definitions/PageVO«ConfigVO»",
          "originalRef": "PageVO«ConfigVO»"
        },
        "message": {
          "type": "string"
        },
        "success": {
          "type": "boolean"
        }
      },
      "title": "Response«PageVO«ConfigVO»»"
    },
...
    "Response«boolean»": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "currentTime": {
          "type": "string",
          "format": "date-time"
        },
        "data": {
          "type": "boolean"
        },
        "message": {
          "type": "string"
        },
        "success": {
          "type": "boolean"
        }
      },
      "title": "Response«boolean»"
    },
    "Response«long»": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "currentTime": {
          "type": "string",
          "format": "date-time"
        },
        "data": {
          "type": "integer",
          "format": "int64"
        },
        "message": {
          "type": "string"
        },
        "success": {
          "type": "boolean"
        }
      },
      "title": "Response«long»"
    },
...
  }
}